JAMES-2890 get rid of the rfc package and use the refined Id
